refactor: 질문 입력 공통 컴포넌트 내부 hooks 분리
